Start work to make TokenEnum contain other useful bits
Make sure that SyntaxErrorNodes always have children, and are not just a leaf node

Convert Token errors into appropriate SyntaxErrorNodes when abstracting

Allow e.g., { + } to be parsed correctly as boxes

Revamped the error handling in the parser, so that unexpected closing brackets and unexpected operators do not eat any unnecessary whitespace. Lot of work and kind of ugly. But maybe error handling has to be ugly