Introduce specialized parselets for better error handling
Move all getPrecedence() logic to inside of parselets, so that Parser class does not have to do any work for looking up preceence. This results in a big speedup now.

Move all handleNotPossible() logic to inside of parselets, this is also a speedup, but this is already an error case, so not that big of a deal.