Fix issue with implicit Times via a large engineering effort.
Parsing a b c should give a single Times node, but what was happening was that the second implicit Times token was not being accounted for properly.

Add a processImplicitTimes() method now that is called in places where it is possible for implicit TImes to appear. This returns the same input token by default. But also returns an implicit Times tokens when necessary.

And since we are in the process of giving more structure to parselets, then also store the Op symbol that is used by operator parselets.