Improve stack shim performance
Use nested tuple instead of AppendTo to improve the top-level expression abstraction.
The old implementation takes O(n) for Push, and Pop, and O(1) for Normal,
while the new implementation takes O(1) for Push, and Pop, and O(n) for Normal.
This accelerates the `abstractTopLevel` function since it contains n pushes and pops and m normalizations of size n/m.
Thus, this change improves it from O(n^2) to O(n), which is efficient enough to remove the $TopLevelExpressionLimit.