Big LineContinuation changes
In preparation for work on a code formatter, I needed to revisit line continuation handling in the parser.

Remove Token`LineContinuation. The actual line continuation string is now always attached to the previous or next token.
Currently, it's always the next token, but that is an implementation detail.

Remove the 3 fake line continuation code points. These are no longer used.

Remove the 2 NextPolicy bits that were related to line continuations. These are no longer used.

Remove UnexpectedLineContinuation FormatIssue. This is no longer used.

Remove isMBLineContinuation() property. This is no longer used.

NOTE: line continuations are NOT currently handled in places where SourceCharacters are used directly: comments and file stringification. This is not relevant for comments, but it is wrong to not handle line continuations during file stringification.