Use std::set for Issues and CodeActions.
This will help with the problem of duplicated issues being reported because of the current way that parsing happens (it is possible to generate messages when peeking ahead and reading the current token).

Setup reporting of embedded tabs in CharacterDecoder. Tabs that occur after a line continuation need to also be reported.

Also rename uses of BackSlash -> Backslash