experiment: Panic instead of generating AbortNode's
These exeptions will automatically be caught and transformed into Failure[..] objects by
the #[export(wstp)] wrapper at the LibraryLink API layer.

This does remove some logic for fixing up the parsing state at the time of the
abort. I'm not completely certain if/when that logic is intended to have an effect.

E.g. one guess might be that the parser is attempting to be resumable, so e.g. an
abort halfway through parsing a large file would preserve the state of the already
parsed tokens so that a subsequent call to resume parsing wouldn't have to re-parse
the first half of the file. However, there doesn't seem to be any of the other logic
require to make that work. So my guess is that the minimal logic present in this
`if <did abort> { .. }` blocks is purely there to push the AbortNode.

Making this change removes many many calls to Parser_tryContinue, which makes it
one step easier to remove the experiment nightly feature(trait_upcasting), which
will allow this project to compile on stable Rust.

This change has no noticable effect on benchmark performance. However, none of
those benchmarks simulate an abort occurring partway through the benchmark.

fixup: Remove AbortNode and Node::Abort(_)