cleanup: Remove unused Parser_tryContinue 2nd parameter (_: ParseletPtr)
Removing this makes it easier to remove the `feature(trait_upcasting)`,
because Parser_tryContinue is used very generically in many places
from both infix and prefix parselets, but it takes an (unused) generic
ParseletPtr argument.

Without trait upcasting, it requires a bit of ceremony to go from a
PrefixParseletPtr or InfixParseletPtr to ParseletPtr. Since this
parameter is unused anyway, it seems easier to just remove it.