cleanup: Remove unused Parser_parseClimb 2nd parameter (Ignored: ParseletPtr)
Like with the previous commit changing Parser_tryContinue, removing
this makes it easier to remove the `feature(trait_upcasting)`,
because Parser_tryClimb is used very generically in many places
from both infix and prefix parselets, but it takes an (unused) generic
ParseletPtr argument.

Without trait upcasting, it requires a bit of ceremony to go from a
PrefixParseletPtr or InfixParseletPtr to ParseletPtr. Since this
parameter is unused anyway, it seems easier to just remove it.