cleanup: Removed unused `ignored: ParseletPtr` parameters from *Parselet member functions
This parameter is essentially the `&self` binding to the parselet statics
stored in the prefixParselets/infixParselets statics. However, since many
parselets have no associated state, there is no need from them to access `self`.

These parameters would be easy to re-add in the future to any parselets
that gain state.