cleanup: Remove unused `ignored: Token` parameter from Parser_parseClimb and Parser_tryContinue
...and also remove `ignored: Token` parameter from parselet functions that are
set as callbacks on Context.f.

This parameter value seems to always be the first token that began a parselet,
but there is no parsing logic that every uses this value -- it's just being
passed down the call stack in circles.

If there is ever a parselet that needs this value, it might be easier to
handle that on a case-by-case basis by using a closure to capture that Token
value in the specific parselet callback that needs it, instead of passing it
to many parselets that don't need it.

This didn't seem to have a statistically significant affect on benchmark
performance.