cleanup: Simplify MyString type
This type was a (&'static str, c_int) tuple, where the c_int was
used as a unique identifier, presumably for fast comparison.

However, this scheme has the disadvantage of making it difficult
to construct new MyString instances dynamically, which is needed
when deserializing WL expressions into Rust structs.

I could have added a lookup table of &str -> MyString instance,
but a check of the places where MyString's are actually compared
and a benchmarking run showed that it would be viable to just
remove the `id` field without a noticable performance regression.