port: Revert 156c63c7 "break TokenEnum into a struct with bit-fields"
This commit was made to the master branch after the Rust port was forked off,
and is not immediately compatible with the Rust representation of the TokenEnum
type.

Its not clear if this change was made in the master branch as an optimization or
a code readability improvement; however based on the nature of the change, it
seems more likely to be a code readability improvement, and therefore less essential
to preserve in the ported version.