cleanup(optimization): Remove `offset` field from Buffer
This simplifies and shrinks the sizes of several types, including:

* Token
* TokenStart
* Buffer
* BufferAndLength
* BorrowedTokenInput

And results in 3-6% performance improvement on the large benchmarks.

In the few places where offset is needed, it can be calculated using
the new `Reader::offset_of(&self, Buffer)` method.

This also makes the construction of Token instances for testing
purposes simpler, because the offset value no longer needs to be
provided so that equality comparisons succeed.

* Removed ByteSpan and related methods