cleanup: Remove `SyntaxErrorNode.src` field
This field is always computed to be exactly equal to
Span::between(..) of the children of the SyntaxErrorNode.

Storing it instead of computing it as-needed on the fly means
this extra space used and duplicate information that could get
out of date with the children node if they're updated.

The tradeoff is computing the source on the fly may have negative
implications for performance, especially when the source of every
node has to be computed recursively for serialization to send
across a WSTP link. With this uncommon node type that seems likely
to be negligible at worst.