cleanup: Avoid intermediate push/pops using NodeStack
When constucting nodes that take a fixed number of arguments,
we can avoid pushing nodes to the NodeStack when we're immediately
going to follow those pushes with a call to reduce_and_climb() that
will just pop them off the stack.

This allows us to construct nodes with greater strongly-typed "structure",
because that structured isn't lost due to "serialization" of nodes in the
flat sequence that is the node stack -- the the new `new2()` methods on
several node types.