cleanup(WIP): Move context_stack to ParserSession from ParseCst
This has a -8-13% performance *regression*, because what was previously
just an access of the `index: usize` field now first requires Box'ing
the ParseCstContext struct, and then doing a downcast::<ParseCstContext>()
every time a reduce_*() operation is done.

However, this is conceptually a step in the right direction because it
means that individual ParseBuilder impls don't have to own the context
stack, which is a core part of the parser logic.

Note that the alternative to using a `dyn Any` for the arbitrary context
data would have been to add an associated type to the ParseBuilder trait.
However, doing so would have made it more difficult to eventually use a
`dyn ParseBuilder` in ParserSession, since traits with associated types
can't be used with `dyn`.

* Store and pass `dyn Any` context data provided by ParseBuilder impl