feature: Support arbitrary ParseBuilder implementations
This makes it possible to directly produce parsed output
of types other than Cst. For example, making the parser directly
product parsed expressions.

The ParseBuilder trait was introduced in a previous commit,
but ParserSession was still hard-coded to always use ParseCst.
This has now been generalized, and now any type that implements
ParseBuilder can be used to build up a parsed output.

This required introducing DynParseBuilder, which is stored as
a `&mut dyn DynParseBuilder` on ParserSession. Using dynamic
dispatch instead of a type parameter on ParserSession was chosen
to avoid the inherent complexity of making ParserSession generic,
and the related complexity of then also needing to make PrefixParselet
and InfixParselet generic, which is especially tricky due to the
precomputed INFIX_PARSELETS and PREFIX_PARSELETS const arrays,
which themselves contain `&dyn InfixParselet` and `&dyn PrefixParselet`
values.