cleanup: Remove Span representation of "unknown" source loc + add Source::Unknown
Span values are constructed when the concrete tokenizer is doing its
work, and it should not be possible for the tokenizer to not know where
it is in the input to product a Span for the current token.

The notion of unknown Span's comes from two places:

1. The Ast, where subexpressions _can_ have unknown source locations
   (e.g. if they're some synthetic expression implied by the literal syntax
   but not actually appearing in the original input).

2. The expression representation of nodes and tokens. If the from_expr.rs
   impl gets an expression constructed by the user, it might not have a Source
   location.

With this refactoring, the new Source::Unknown variant of the more general
Source type can instead be used to represent unknown source spans in the
two cases listed above.