feature: Add new enum BoxPosition { At(..), Spanning(..) }
This is needed so that box positions that end in a Span[start, finish] range
can be represented. Which in turn is ultimately needed so that I can
remove the `src` field of most remaining `Cst` node types and still have
the computed sources return the correct value, which will ocassionaly require
intermediate sources with this new Spanning {..} position value.