test: Add test for aggregate/abstract of Token`Boxes`MultiWhitespace
This token is currently special in that it currently does not
satisfy isTrivia(), but it is removed by aggregate_cst().

This may just be an artifact of the historical reality that the
compiled parser implementation did not have to handle nodes
parsed from boxes, and Token`Boxes`MultiWhitespace should be
made to satisfy isTrivia().