test: Add exhaustive tests for ternary ~ valid, invalid, and unusual precedence cases
This is in preparation to _remove_ CodeParser functionality for
exactly matching the *quirky and erroneous* Kernel parsing behavior
for ~ syntax.

This is motivated by an ongoing refactoring in which the current
parser implemention's assumption that the parsed value is a Cst
is violated, which makes it onerous to continue supporting the
quirky ~ parsing behavior via the special
top_non_trivia_node_is_tilde() method.

CodeParser will instead match the FrontEnd parsing behavior for ~,
and emit syntax errors in unusual (and presumably quite rare) cases
it (and the Kernel) currently accept.

1. All previously supported syntax will now be a syntax error, NOT
   a different --valid-- parse. This change would be unjustifiable
   if it silently changed how ~ parsed into an expr. But this change
   is to reject previously accepted syntax with a hard syntax error,
   NOT to parse it into some subtley different expression.

   This will be potentially inconvenient for the presumably very small
   subset users who see new syntax errors in old code. However, the cases
   where this old quirky parsing was accepted were unusual and highly
   inconsistent -- e.g. "a~f;~b" was previously parsed as `(f;)[a, b]`

2. The scope of potentialy effected users is limited because the previous
   accepted syntax was NOT accepted by the FE parser. So e.g. any code
   originally authored in the FE and then pasted into a package .m/.wl
   file could not contain this erroneous syntax.