bugfix/feature: Perform aggregate[..] in pure WL
This avoids the need to convert node epressions going from
WL => Rust => WL just to delete trivia nodes, which is a
very simple transformation.

The primary motivation for doing this was that aggregate[..]
is sometimes called on non-ContainerNode root values (in particular,
by CodeInspector), which is a problem because the implementation
of Source::from_expr() needs to know what the structure of the
Source field is to correctly do the expr => Rust struct parsing.

That problem is exacerbated the the refactoring to remove the `src`
field from most node types, which has the consequence that the
Source value for non-token nodes has to be recomputed during the
Rust => WL expression conversion -- which is impossible if the
original Source field couldn't be parsed correctly.

So easier to just move the implementation of aggregate[..] from
Rust to WL.

abstract[..] doesn't have this problem because it is always done
on complete ContainerNode[tag, ..] values, where we can guide the
Source::from_expr() implementation based on the `tag`.