cleanup: Add ParseBuilder::{TriviaAccumulator, TriviaHandle} abstraction
This allows the ParseBuilder implementation to decide
how--and whether--to record and process trivia tokens.

This should enable certain parse implementations that
don't care about the position of trivia in the input to
implement the new associated types as empty types, avoiding
the parser ever storing trivia tokens beyond the minimal
processing needed by the underlying Tokenizer calls to
read past them.